Поменял тип счётчиков заполненности октетов с short на int, т.к. разницы по расходу памяти не будет. Эти 2 байта "съест" выравнивание.

+++ src/main/java/pavel_kuzenkov/com/github/unique_IP_counter/ip_address_entity/LastOctet.java
@@ -23,7 +23,7 @@ class LastOctet implements Octet {
     /**
      * Счетчик заполненности массива с уникальными адресами.
      */
-    private short fillCounter = 0;
+    private int fillCounter = 0;
 
     /**
      * Проверка данного октета на заполненность.
